fixed small bugs
1. fixed conversion int->double that did not work when the number
of generations was large, in the calculation of concordance factors.
2. alpha forced to be +Inf with the independence prior: fixed so
that there is no warning when compiling.
3. fixed conversion int->double to use log in logUB

+++ src/bucky.h
@@ -36,8 +36,8 @@ double logUB(unsigned int n){
   }
   if (n<4) return 0.0;
   double answer = 0.0;
-  for (int i=4;i<n+1;i++)
-    answer += log(2*i - 5);
+  for (unsigned int i=4;i<n+1;i++)
+    answer += log((double)(2*i - 5));
   return answer;
 }
